requirements.txt
ssh2-python